fixup! feat(@trezor/suite-web-landing): remove next.js & use custom build pipeline
