+++ packages/suite-web-landing/src/static/index.html
@@ -15,9 +15,15 @@
             rel="apple-touch-icon"
             href="<%= assetPrefix %>/static/images/icons/favicon/favicon.png"
         />
-        <script src="<%= assetPrefix %>/static/browser-detection/index.js"></script>
+        <% if(isOnionLocation) { %>
+            <meta
+                httpEquiv="onion-location"
+                content="http://suite.trezoriovpjcahpzkrewelclulmszwbqpzmzgub37gbcjlvluxtruqad.onion/web"
+            />
+        <% } %>
     </head>
     <body>
+        <div id="app"></div>
         <noscript>This website requires JavaScript to run.</noscript>
     </body>
 </html>
