+++ packages/suite-web-landing/src/support/index.ts
@@ -1,0 +1,5 @@
+import ReactDOM from 'react-dom';
+import './index.css';
+import Main from '../Main';
+
+ReactDOM.render(Main, document.getElementById('app'));
