+++ packages/suite-web/src/static/index.html
@@ -3,6 +3,9 @@
     <head>
         <meta name="google" content="notranslate" />
         <meta charset="utf-8">
+        <meta httpEquiv="Pragma" content="no-cache" />
+        <meta httpEquiv="cache-control" content="no-cache" />
+        <meta httpEquiv="expires" content="-1" />
         <title>Trezor Suite</title>
 
         <link
